+++ lib/apiScenario/apiScenarioSchema.ts
@@ -7,8 +7,8 @@ export const ApiScenarioDefinition: Schema & {
   properties: {
     scope: {
       type: "string",
-      enum: ["ResourceGroup", "None"],
-      default: "None",
+      enum: ["ResourceGroup", "Subscription", "Tenant", "None"],
+      default: "ResourceGroup",
     },
     variables: {
       $ref: "#/definitions/Variables",
@@ -338,18 +338,29 @@ export const ApiScenarioDefinition: Schema & {
           additionalProperties: false,
           patternProperties: {
             "^([0-9]{3})$": {
-              type: "object",
-              properties: {
-                headers: {
+              oneOf: [
+                {
                   type: "object",
-                  additionalProperties: {
-                    type: "string",
+                  properties: {
+                    headers: {
+                      type: "object",
+                      additionalProperties: {
+                        type: "string",
+                      },
+                    },
+                    body: {
+                      type: ["object", "number", "array", "integer", "string", "boolean", "null"],
+                    },
+                  },
+                  additionalProperties: false,
+                },
+                {
+                  type: "array",
+                  items: {
+                    $ref: "#/definitions/JsonTestOp",
                   },
                 },
-                body: {
-                  type: ["object", "number", "array", "integer", "string", "boolean", "null"],
-                },
-              },
+              ],
             },
           },
         },
@@ -472,9 +483,6 @@ export const ApiScenarioDefinition: Schema & {
         {
           $ref: "#/definitions/JsonPatchOpMove",
         },
-        {
-          $ref: "#/definitions/JsonPatchOpTest",
-        },
       ],
     },
     JsonPatchOpAdd: {
@@ -535,16 +543,40 @@ export const ApiScenarioDefinition: Schema & {
       },
       additionalProperties: false,
     },
-    JsonPatchOpTest: {
+    JsonTestOp: {
       type: "object",
-      required: ["test", "value"],
+      required: ["test"],
       properties: {
         test: {
           $ref: "#/definitions/JsonPointer",
         },
-        value: {},
       },
-      additionalProperties: false,
+      allOf: [
+        {
+          oneOf: [
+            {
+              type: "object",
+              required: ["value"],
+              properties: {
+                test: {},
+                value: {},
+              },
+              additionalProperties: false,
+            },
+            {
+              type: "object",
+              required: ["expression"],
+              properties: {
+                test: {},
+                expression: {
+                  type: "string",
+                },
+              },
+              additionalProperties: false,
+            },
+          ],
+        },
+      ],
     },
   },
 };