+++ lib/util/utils.ts
@@ -323,9 +323,7 @@ const parse = (pointer: string) => {
 };
 
 export function getProviderFromFilePath(pathStr: string): string | undefined {
-  const resourceProviderPattern: RegExp = new RegExp(
-    `^[A-Z][a-z0-9]+(\.([A-Z]{1,3}[a-z0-9]+)+[A-Z]{0,2})+$`
-  );
+  const resourceProviderPattern: RegExp = /[A-Z][a-z0-9]+(\.([A-Z]{1,3}[a-z0-9]+)+[A-Z]{0,2})+/g;
   const words = pathStr.split(/\\|\//gi);
   for (const it of words) {
     if (resourceProviderPattern.test(it)) {
@@ -974,4 +972,52 @@ export const getRandomString = (length?: number) => {
     .next()
     .value.toString(36)
     .slice(0 - (length ?? 6));
+};
+
+export const findPathsToKey = (options: {
+  key: string;
+  obj: any;
+  pathToKey?: string;
+}): string[] => {
+  const results = [];
+  (function findKey({ key, obj, pathToKey }) {
+    const oldPath = `${pathToKey ? pathToKey : ""}`;
+    if (obj && obj.hasOwnProperty(key)) {
+      results.push(`${oldPath}.${key}`);
+      return;
+    }
+    if (obj !== null && typeof obj === "object" && !Array.isArray(obj)) {
+      for (const k in obj) {
+        if (obj.hasOwnProperty(k)) {
+          if (Array.isArray(obj[k])) {
+            for (let j = 0; j < obj[k].length; j++) {
+              findKey({
+                obj: obj[k][j],
+                key,
+                pathToKey: `${oldPath}${k}['${j}']`,
+              });
+            }
+          }
+          if (obj[k] !== null && typeof obj[k] === "object") {
+            findKey({
+              obj: obj[k],
+              key,
+              pathToKey: /[\*|\{|\[|\}|\}|\,|\.]/.test(k)
+                ? `${oldPath}['${k}']`
+                : `${oldPath}.${k}`,
+            });
+          }
+        }
+      }
+    }
+  })(options);
+
+  return results;
+};
+
+export const findPathToValue = (arr: string[], obj: any, value: string) => {
+  return arr.reduce((pre: string[], cur: string) => {
+    lodash.get(obj, cur.substr(1)) === value && pre.push(cur);
+    return pre;
+  }, []);
 };